k8s: Move useNodeCIDR test to appropriate file
This commit makes no functional changes. The function `useNodeCIDR` is
defined and used inside pkg/k8s/init.go and therefore belongs inside
this new file.
